+++ cherimoya/__init__.py
@@ -4,5 +4,9 @@
 from .cherimoya import Cherimoya
 from .cherimoya import EMA
 from .cheri import CheriBlock
+from .wrappers import ControlWrapper
+from .wrappers import ProfileWrapper
+from .wrappers import LogCountWrapper
+from .wrappers import ExpectedCountsWrapper
 
 __version__ = '0.1.0'